atlast_ekscopy1/requirements.txt
aws-cdk-lib==2.123.0
#2.83.1
constructs>=10.0.0,<11.0.0
